Polish to chrome_remote_control to reduce flake and improve modularity.
- Make android work
- Overhaul tools/gpu to be more component-y
- Fix error where pexpect missing breaks all browser discovery
- Use page.navigate and get rid of the strange timeout logic in InspectorBackend
- Make browser selection explicit. You have to select browsers explicitly.
- Add --repeat-count to run_tests
- Move sleep-after-each-load to gpu_tools.url_test
- Make timeouts explicit
- Fix handling of optparse defaults and counters
- Add is_content_shell property on backends
- Set up content-shell command line on android
- Override user-data-dir only for non-content-shell backends
- If browser fails to die on sigterm, send it a sigkill
- Move gpu tools browser creation to toplevel test
- Make non-jsonizable evaluate test independent of window for perf reasons
- Add runtime.execute to avoid hacks of runtime.evaluate

TBR=dtu@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10912078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@154754 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8b46deaa1f0ab3794fa08f08f314c481584e591a

+++ tools/chrome_remote_control/chrome_remote_control/desktop_browser_backend.py
@@ -11,6 +11,7 @@ import browser_backend
 import browser_finder
 import inspector_backend
 import tab
+import util
 
 DEFAULT_PORT = 9273
 
@@ -18,8 +19,8 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
   """The backend for controlling a locally-executed browser instance, on Linux,
   Mac or Windows.
   """
-  def __init__(self, options, executable):
-    super(DesktopBrowserBackend, self).__init__()
+  def __init__(self, options, executable, is_content_shell):
+    super(DesktopBrowserBackend, self).__init__(is_content_shell)
 
     # Initialize fields so that an explosion during init doesn't break in Close.
     self._proc = None
@@ -30,12 +31,12 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
     if not self._executable:
       raise Exception("Cannot create browser, no executable found!")
 
-    self._tmpdir = tempfile.mkdtemp()
     self._port = DEFAULT_PORT
     args = [self._executable,
             "--no-first-run",
             "--remote-debugging-port=%i" % self._port]
     if not options.dont_override_profile:
+      self._tmpdir = tempfile.mkdtemp()
       args.append("--user-data-dir=%s" % self._tmpdir)
     args.extend(options.extra_browser_args)
     if options.hide_stdout:
@@ -60,12 +61,34 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
 
   def Close(self):
     if self._proc:
-      self._proc.terminate()
-      self._proc.wait()
-      self._proc = None
 
-    if os.path.exists(self._tmpdir):
+      def IsClosed():
+        if not self._proc:
+          return True
+        return self._proc.poll() != None
+
+      # Try to politely shutdown, first.
+      self._proc.terminate()
+      try:
+        util.WaitFor(IsClosed, timeout=1)
+        self._proc = None
+      except util.TimeoutException:
+        pass
+
+      # Kill it.
+      if not IsClosed():
+        self._proc.kill()
+        try:
+          util.WaitFor(IsClosed, timeout=5)
+          self._proc = None
+        except util.TimeoutException:
+          self._proc = None
+          raise Exception("Could not shutdown the browser.")
+
+    if self._tmpdir and os.path.exists(self._tmpdir):
       shutil.rmtree(self._tmpdir, ignore_errors=True)
+      self._tmpdir = None
 
     if self._devnull:
       self._devnull.close()
+      self._devnull = None
