+++ tools/gpu/gpu_tools/file_server.py
@@ -6,6 +6,8 @@ import subprocess
 
 class FileServer(object):
   def __init__(self, path, port=8000):
+    self._server = None
+    self._devnull = None
     assert os.path.exists(path)
     if os.path.isdir(path):
       self._path = path
@@ -13,16 +15,30 @@ class FileServer(object):
       self._path = os.path.dirname(path)
     self._port = port
 
+
   def __enter__(self):
+    self._devnull = open(os.devnull, 'w')
     self._server = subprocess.Popen(
         ['python', '-m', 'SimpleHTTPServer', str(self._port)],
-        cwd=self._path)
+        cwd=self._path,
+        stdout=self._devnull, stderr=self._devnull)
+    return self
+
+  @property
+  def url(self):
     return 'http://localhost:%d' % self._port
 
   def __exit__(self, *args):
-    self._server.kill()
-    self._server = None
+    if self._server:
+      self._server.kill()
+      self._server = None
+    if self._devnull:
+      self._devnull.close()
+      self._devnull = None
 
   def __del__(self):
     if self._server:
       self._server.kill()
+    if self._devnull:
+      self._devnull.close()
+      self._devnull = None
